flatten rev meta
