+++ stan/math/mix/meta.hpp
@@ -8,8 +8,8 @@
 #include <stan/math/fwd/meta/partials_type.hpp>
 
 #include <stan/math/rev/core.hpp>
-#include <stan/math/rev/scal/meta/is_var.hpp>
-#include <stan/math/rev/scal/meta/partials_type.hpp>
+#include <stan/math/rev/meta/is_var.hpp>
+#include <stan/math/rev/meta/partials_type.hpp>
 
 #include <stan/math/fwd/meta.hpp>
 #include <stan/math/rev/meta.hpp>
