+++ stan/math/rev/meta.hpp
@@ -1,10 +1,10 @@
 #ifndef STAN_MATH_REV_META_HPP
 #define STAN_MATH_REV_META_HPP
 
-#include <stan/math/rev/scal/meta/is_var.hpp>
-#include <stan/math/rev/scal/meta/partials_type.hpp>
-#include <stan/math/rev/mat/meta/operands_and_partials.hpp>
+#include <stan/math/rev/meta/is_var.hpp>
+#include <stan/math/rev/meta/partials_type.hpp>
+#include <stan/math/rev/meta/operands_and_partials.hpp>
 
-#include <stan/math/rev/scal/meta/operands_and_partials.hpp>
+#include <stan/math/rev/meta/operands_and_partials.hpp>
 
 #endif
