+++ stan/math/rev/meta/is_var.hpp
@@ -1,5 +1,5 @@
-#ifndef STAN_MATH_REV_SCAL_META_IS_VAR_HPP
-#define STAN_MATH_REV_SCAL_META_IS_VAR_HPP
+#ifndef STAN_MATH_REV_META_IS_VAR_HPP
+#define STAN_MATH_REV_META_IS_VAR_HPP
 
 #include <stan/math/prim/scal/meta/is_var.hpp>
 #include <stan/math/rev/core.hpp>
