+++ stan/math/rev/meta/operands_and_partials.hpp
@@ -1,5 +1,5 @@
-#ifndef STAN_MATH_REV_SCAL_META_OPERANDS_AND_PARTIALS_HPP
-#define STAN_MATH_REV_SCAL_META_OPERANDS_AND_PARTIALS_HPP
+#ifndef STAN_MATH_REV_META_OPERANDS_AND_PARTIALS_HPP
+#define STAN_MATH_REV_META_OPERANDS_AND_PARTIALS_HPP
 
 #include <stan/math/rev/core/chainablestack.hpp>
 #include <stan/math/rev/core/precomputed_gradients.hpp>
@@ -8,6 +8,10 @@
 #include <stan/math/prim/scal/meta/broadcast_array.hpp>
 #include <stan/math/prim/scal/meta/operands_and_partials.hpp>
 #include <stan/math/prim/scal/meta/is_vector_like.hpp>
+#include <stan/math/prim/scal/meta/likely.hpp>
+#include <stan/math/rev/mat/fun/typedefs.hpp>
+#include <stan/math/prim/arr/meta/length.hpp>
+#include <vector>
 
 namespace stan {
 namespace math {
@@ -125,6 +129,155 @@ class operands_and_partials<Op1, Op2, Op3, Op4, Op5, var> {
     return var(new precomputed_gradients_vari(value, size, varis, partials));
   }
 };
+
+namespace internal {
+// Vectorized Univariate
+template <>
+class ops_partials_edge<double, std::vector<var>> {
+ public:
+  using Op = std::vector<var>;
+  using partials_t = Eigen::VectorXd;
+  partials_t partials_;                       // For univariate use-cases
+  broadcast_array<partials_t> partials_vec_;  // For multivariate
+  explicit ops_partials_edge(const Op& op)
+      : partials_(partials_t::Zero(op.size())),
+        partials_vec_(partials_),
+        operands_(op) {}
+
+ private:
+  template <typename, typename, typename, typename, typename, typename>
+  friend class stan::math::operands_and_partials;
+  const Op& operands_;
+
+  void dump_partials(double* partials) {
+    for (int i = 0; i < this->partials_.size(); ++i) {
+      partials[i] = this->partials_[i];
+    }
+  }
+  void dump_operands(vari** varis) {
+    for (size_t i = 0; i < this->operands_.size(); ++i) {
+      varis[i] = this->operands_[i].vi_;
+    }
+  }
+  int size() { return this->operands_.size(); }
+};
+
+template <int R, int C>
+class ops_partials_edge<double, Eigen::Matrix<var, R, C>> {
+ public:
+  using Op = Eigen::Matrix<var, R, C>;
+  using partials_t = Eigen::Matrix<double, R, C>;
+  partials_t partials_;                       // For univariate use-cases
+  broadcast_array<partials_t> partials_vec_;  // For multivariate
+  explicit ops_partials_edge(const Op& ops)
+      : partials_(partials_t::Zero(ops.rows(), ops.cols())),
+        partials_vec_(partials_),
+        operands_(ops) {}
+
+ private:
+  template <typename, typename, typename, typename, typename, typename>
+  friend class stan::math::operands_and_partials;
+  const Op& operands_;
+
+  void dump_operands(vari** varis) {
+    for (int i = 0; i < this->operands_.size(); ++i) {
+      varis[i] = this->operands_(i).vi_;
+    }
+  }
+  void dump_partials(double* partials) {
+    for (int i = 0; i < this->partials_.size(); ++i) {
+      partials[i] = this->partials_(i);
+    }
+  }
+  int size() { return this->operands_.size(); }
+};
+
+// SPECIALIZATIONS FOR MULTIVARIATE VECTORIZATIONS
+// (i.e. nested containers)
+template <int R, int C>
+class ops_partials_edge<double, std::vector<Eigen::Matrix<var, R, C>>> {
+ public:
+  using Op = std::vector<Eigen::Matrix<var, R, C>>;
+  using partial_t = Eigen::Matrix<double, -1, -1>;
+  std::vector<partial_t> partials_vec_;
+  explicit ops_partials_edge(const Op& ops)
+      : partials_vec_(ops.size()), operands_(ops) {
+    for (size_t i = 0; i < ops.size(); ++i) {
+      partials_vec_[i] = partial_t::Zero(ops[i].rows(), ops[i].cols());
+    }
+  }
+
+ private:
+  template <typename, typename, typename, typename, typename, typename>
+  friend class stan::math::operands_and_partials;
+  const Op& operands_;
+
+  void dump_partials(double* partials) {
+    int p_i = 0;
+    for (size_t i = 0; i < this->partials_vec_.size(); ++i) {
+      for (int j = 0; j < this->partials_vec_[i].size(); ++j, ++p_i) {
+        partials[p_i] = this->partials_vec_[i](j);
+      }
+    }
+  }
+  void dump_operands(vari** varis) {
+    int p_i = 0;
+    for (size_t i = 0; i < this->operands_.size(); ++i) {
+      for (int j = 0; j < this->operands_[i].size(); ++j, ++p_i) {
+        varis[p_i] = this->operands_[i](j).vi_;
+      }
+    }
+  }
+  int size() {
+    if (unlikely(this->operands_.size() == 0)) {
+      return 0;
+    }
+    return this->operands_.size() * this->operands_[0].size();
+  }
+};
+
+template <>
+class ops_partials_edge<double, std::vector<std::vector<var>>> {
+ public:
+  using Op = std::vector<std::vector<var>>;
+  using partial_t = std::vector<double>;
+  std::vector<partial_t> partials_vec_;
+  explicit ops_partials_edge(const Op& ops)
+      : partials_vec_(length(ops)), operands_(ops) {
+    for (size_t i = 0; i < length(ops); ++i) {
+      partials_vec_[i] = partial_t(length(ops[i]), 0.0);
+    }
+  }
+
+ private:
+  template <typename, typename, typename, typename, typename, typename>
+  friend class stan::math::operands_and_partials;
+  const Op& operands_;
+
+  void dump_partials(double* partials) {
+    int p_i = 0;
+    for (size_t i = 0; i < this->partials_vec_.size(); ++i) {
+      for (size_t j = 0; j < this->partials_vec_[i].size(); ++j, ++p_i) {
+        partials[p_i] = this->partials_vec_[i][j];
+      }
+    }
+  }
+  void dump_operands(vari** varis) {
+    int p_i = 0;
+    for (size_t i = 0; i < this->operands_.size(); ++i) {
+      for (size_t j = 0; j < this->operands_[i].size(); ++j, ++p_i) {
+        varis[p_i] = this->operands_[i][j].vi_;
+      }
+    }
+  }
+  int size() {
+    if (unlikely(this->operands_.size() == 0)) {
+      return 0;
+    }
+    return this->operands_.size() * this->operands_[0].size();
+  }
+};
+}  // namespace internal
 }  // namespace math
 }  // namespace stan
 #endif
